Version 10.0.2.1
DNS Techpack 10.0.2.0 - Full changelog:
https://github.com/DNSTechpack/DNS10-MC1.10.2/commits/master

Updated: (7)
- Chisel
- Ender IO
- JourneyMap
- Reborn Core
- Storage Drawers
- Techreborn
- ThutCore

Added: (0)

Notes:
- Fixed the random crash caused by Ender IO piracy detection on startup
- Thanks Henry :D
- Fixed JourneyMap button causing client crash
- Set BoP worldtype to default

I am aware of GUI crashes with Refined Storage and Draconic Evolution! I
am awaiting fixes and comments on the bug reports.

Refinsed Storage is supposed to be fixed, but the fixed version is not
released yet. This will be coming for 10.0.2.2 provided that it is
released soon.